requirements.txt
streamlit
google-generativeai
pdf2image
python-dotenv
